backend/requirements.txt
fastapi
uvicorn
groq
python-multipart
openai-whisper
opencv-python
deepface
torch
python-dotenv